fix: update ER diagram  neo look test case
on-behalf-of: @Mermaid-Chart <hello@mermaidchart.com>

+++ cypress/integration/rendering/erDiagram-neo-look.spec.ts
@@ -138,7 +138,7 @@ looks.forEach((look) => {
         string[] tags
         varchar(255) description
         type~T~ genericType
-        decimal(10,2) price
+        float price
     }
 `;
         imgSnapshotTest(erCode, { look, theme });
